Started working on the view of AngularJS.
Not sure if I'll go with a approach like Rails, with multiple routes, or a more simple view with a "all in one"
approach. I'm also stuck on a problem that I need some information prompted on another sub-generator to be used
inside another sub-generator. I already asked the user about the fields present on the model, doesn't make any
sense ask this again.

I'm taking advantage of the hookFor system to create some files on the Angular side. The problem is that i need
to override those methods because the action on the generator-angular don't let me to just create the route on
app.js for example. Because of that I need to run the action on generator-angular first, then afterwards I run
my method with a force flag. Weird, but words. Let see if I can found a solution for that.

I'm looking forward to improve the USAGE description.